refactor: MemberFixture 임시 네이밍 변경 #114
